cheri: wireless: Apply the multicast.cocci script to update the ABI
Apply the multicast.cocci script to update the function definitions
for the prepare_multicast and configure_filter scripts.

In order to catch prototypes in header files and implementations
that are in a different file first search for all _directories_
continaing at least one file that contains the strings prepare_multicast
and/or configure_filter.

Then run coccinelle on each directory <DIR> like this:
	spatch \
		--sp-file scripts/coccinelle/cheri/multicast.cocci \
		--macro-file scripts/coccinelle/cheri/multicast.h \
		<DIR>/*.[ch]

+++ drivers/net/wireless/virtual/mac80211_hwsim.c
@@ -2467,7 +2467,8 @@ static int mac80211_hwsim_config(struct ieee80211_hw *hw, u32 changed)
 
 static void mac80211_hwsim_configure_filter(struct ieee80211_hw *hw,
 					    unsigned int changed_flags,
-					    unsigned int *total_flags,u64 multicast)
+					    unsigned int *total_flags,
+					    uintptr_t multicast)
 {
 	struct mac80211_hwsim_data *data = hw->priv;
